// src/app/types/Employee.ts
export interface Employee {
    /**
     * 
     */
     userid?: string;

     /**
      * 
      */
     employeeid?: string;
 
     /**
      * 
      */
     password?: string;
 
     /**
      * 
      */
     firstname?: string;
 
     /**
      * 
      */
     middleinitial?: string;
 
     /**
      * 
      */
     lastname?: string;
 
     /**
      * 
      */
     gender?: string;
 
     /**
      * 
      */
     dateofbirth?: string;
 
     /**
      * 
      */
     email?: string;
 
     /**
      * 
      */
     streetname1?: string;
 
     /**
      * 
      */
     city?: string;
 
     /**
      * 
      */
     state?: string;
 
     /**
      * 
      */
     zipcode?: string;
 
     /**
      * 
      */
     country?: string;
 
     /**
      * 
      */
     streetname2?: string;
 
     /**
      * 
      */
     position?: string;
 
     /**
      * 
      */
     homephone?: string;
 
     /**
      * 
      */
     mobilephone?: string
 
     /**
      * 
      */
     workphone?: string;
 
     /**
      * 
      */
     startdate?: string;
 
     /**
      * 
      */
     enddate?: string;
}